Release 2.4.1f (#167)
* Init version 2.4.1f


* Connector type : fix bug for php 7.2
File connector : fix bug to close the file opened via ssh protocol


* Compatibility php7.2 : Fix default controller


* File connector : fix bug to close the file when we get an error during data reading


* Fix bug : Do not remove specific rule param when editing a rule (eg : filename for file connector).


* PHP 7.2 : update manolog version to 3.1.2


* PHP 7.2 change count to !empty


* Php 7.2 fix confitions

+++ src/Myddleware/RegleBundle/Controller/DefaultController.php
@@ -61,8 +61,10 @@ class DefaultControllerCore extends Controller
     protected $em;
     // Connexion direct bdd (utilisé pour créer les tables Z sans doctrine
     protected $connection;
-
-    protected function getInstanceBdd()
+	// Standard rule param list to avoird to delete specific rule param (eg : filename for file connector)
+	protected $standardRuleParam = array('datereference','bidirectional','fieldId','mode','duplicate_fields','limit','delete');
+    
+	protected function getInstanceBdd()
     {
         if (empty($this->em)) {
             $this->em = $this->getDoctrine()->getManager();
@@ -203,7 +205,7 @@ class DefaultControllerCore extends Controller
             $rule = $rule[0];
 
             // si je supprime une règle qui ne m'appartient pas alors redirection
-            if (count($rule) < 1) {
+            if (empty($rule)) {
                 return $this->redirect($this->generateUrl('regle_list'));
             }
 
@@ -1207,9 +1209,10 @@ class DefaultControllerCore extends Controller
                             // Transformation
                             $r['after'][$name_fields_target] = $doc->getTransformValue($source['values'], $target_fields);
 
-                            if (empty($k['champs']))
+							$k['fields'] = array();
+                            if (empty($k['champs'])) {
                                 $k['fields']['Formula'] = ((isset($k['formule'][0]) ? $k['formule'][0] : ''));
-                            else if (count($k['champs']) > 0) {
+                            } else {
                                 foreach ($k['champs'] as $fields) {
                                     // Fields couldn't be return. For example Magento return only field not empty
                                     if (!empty($source['values'][$fields])) {
@@ -1218,8 +1221,6 @@ class DefaultControllerCore extends Controller
                                         $k['fields'][$fields] = '';
                                     }
                                 }
-                            } else {
-                                $k['fields'] = array();
                             }
 
                             $tab_simulation[] = array(
@@ -1959,8 +1960,10 @@ class DefaultControllerCore extends Controller
                         if ($ruleParam->getName() == 'datereference') {
                             $date_reference = $ruleParam->getValue();
                         }
-                        $this->em->remove($ruleParam);
-                        $this->em->flush();
+						if (in_array($ruleParam->getName(), $this->standardRuleParam)) {
+							$this->em->remove($ruleParam);
+							$this->em->flush();
+						}
                     }
                 }
             } // Create mode
@@ -2109,8 +2112,7 @@ class DefaultControllerCore extends Controller
             }
 
             //------------------------------- RuleFilter ------------------------
-
-            if (count($request->request->get('filter')) > 0) {
+            if (!empty($request->request->get('filter'))) {
                 foreach ($request->request->get('filter') as $filter) {
                     $oneRuleFilter = new RuleFilter();
                     $oneRuleFilter->setTarget($filter['target']);
@@ -2274,7 +2276,7 @@ class DefaultControllerCore extends Controller
         $countTransferRule = array();
         $i = 1;
         $values = $home->countTransferRule($permission->isAdmin($this->getUser()->getId()), $this->getUser()->getId());
-        if (count($values) > 0) {
+        if (!empty($values)) {
             foreach ($values as $field => $value) {
                 if ($i == 1) {
                     $countTransferRule[] = array('test', 'test2');
@@ -2458,7 +2460,7 @@ class DefaultControllerCore extends Controller
         $solutionSource = $this->em->getRepository('RegleBundle:Solution')
             ->solutionConnector('source', $permission->isAdmin($this->getUser()->getId()), $this->getUser()->getId());
 
-        if (count($solutionSource) > 0) {
+        if (!empty($solutionSource)) {
             foreach ($solutionSource as $s) {
                 $source[] = $s->getName();
             }
@@ -2469,7 +2471,7 @@ class DefaultControllerCore extends Controller
         $solutionTarget = $this->em->getRepository('RegleBundle:Solution')
             ->solutionConnector('target', $permission->isAdmin($this->getUser()->getId()), $this->getUser()->getId());
 
-        if (count($solutionTarget) > 0) {
+        if (!empty($solutionTarget)) {
             foreach ($solutionTarget as $t) {
                 $target[] = $t->getName();
             }
